DIO_Codes_Intermediates/AplicadoTestesComJUnit/ComMaven/AplicandoTeste/AplicandoTestes/src/main/java/com/one/innovation/digital/Pessoa.java
package com.one.innovation.digital;

import java.time.LocalDate;
import java.time.temporal.ChronoUnit;

public class Pessoa {
    private String nome;
    private LocalDate dataNasc;

    public Pessoa(String nome, LocalDate dataNasc) {
        this.nome = nome;
        this.dataNasc = dataNasc;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public LocalDate getDataNasc() {
        return dataNasc;
    }

    public void setDataNasc(LocalDate dataNasc) {
        this.dataNasc = dataNasc;
    }

    public int getIdade() {
        return (int) ChronoUnit.YEARS.between(this.dataNasc, LocalDate.now());
    }

    public boolean maiorIdade(){return getIdade() >=18;}
}
